pyproject.toml
[project]
name = "taskboard"
version = "0.1.0"
description = "Taskboard infrastructure"
requires-python = ">=3.14.2"
dependencies = [
    "python-dotenv>=1.2.1",
    "fastapi>=0.135.1",
    "uvicorn>=0.41.0",
    "sqlalchemy==2.0.48",
    "psycopg2-binary>=2.9",
    "alembic>=1.13",
    "python-dotenv>=1.0",
    "aws-cdk-lib>=2.0",
    "constructs>=10.0",
    "ruff>=0.15.5",
]

